Add files via upload
This change removes the Recorder from the Wisp file and puts it into a separate audio.py file.
