+++ Wisp.py
@@ -9,6 +9,7 @@ from pydub import AudioSegment
 from faster_whisper import WhisperModel
 import os
 from lang import get_text as tx
+from audio import Recorder, RecordingFile
 
 class MainWindow(QWidget):
     def __init__(self, *args, **kwargs):
@@ -290,63 +291,7 @@ class MainWindow(QWidget):
 
         self.transcription_edit.setText(f"{tx('Transcription_message_1')}'{filename_stem}.txt' {tx('Transcription_message_2')}")
 
-class Recorder(object):
-    # A recorder class for recording audio to a WAV file. Records in mono by default.
 
-    def __init__(self, channels=1, rate=44100, frames_per_buffer=1024):
-        self.channels = channels
-        self.rate = rate
-        self.frames_per_buffer = frames_per_buffer
-
-    def open(self, fname, mode='wb'):
-        return RecordingFile(fname, mode, self.channels, self.rate,
-                            self.frames_per_buffer)
-
-class RecordingFile(object):
-    def __init__(self, fname, mode, channels, 
-                rate, frames_per_buffer):
-        self.fname = fname
-        self.mode = mode
-        self.channels = channels
-        self.rate = rate
-        self.frames_per_buffer = frames_per_buffer
-        self._pa = pyaudio.PyAudio()
-        self.wavefile = self._prepare_file(self.fname, self.mode)
-        self._stream = None
-
-    def start_recording(self):
-        # Use a stream with a callback in non-blocking mode
-        self._stream = self._pa.open(format=pyaudio.paInt16,
-                                            channels=self.channels,
-                                            rate=self.rate,
-                                            input=True,
-                                            frames_per_buffer=self.frames_per_buffer,
-                                            stream_callback=self.get_callback())
-        self._stream.start_stream()
-        return self
-
-    def stop_recording(self):
-        self._stream.stop_stream()
-        return self
-
-    def get_callback(self):
-        def callback(in_data, frame_count, time_info, status):
-            self.wavefile.writeframes(in_data)
-            return in_data, pyaudio.paContinue
-        return callback
-
-
-    def close(self):
-        self._stream.close()
-        self._pa.terminate()
-        self.wavefile.close()
-
-    def _prepare_file(self, fname, mode='wb'):
-        wavefile = wave.open(fname, mode)
-        wavefile.setnchannels(self.channels)
-        wavefile.setsampwidth(self._pa.get_sample_size(pyaudio.paInt16))
-        wavefile.setframerate(self.rate)
-        return wavefile
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
